Make the init function empty
